requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
Werkzeug==2.3.7